ENH: Extract the slice-projection math into the shared Lib
Hoist the slice-view projection / distance fade / signed side tint / hard
presence-cutoff math (ADR-0038 §Context) into
SlicerLiverInteractionLib.SlicePointProjection, the pure-math half of the
ADR-0038 slice base.  VascularTerritoriesLib.TerritorySliceProjection now
re-exports it (behaviour-preserving, ADR-0038 [review]); the resection slice
pipeline delegates to it in a later commit.

Also fix the shared-Lib skeleton REPO_ROOT depth (parents[3]) so the base
tests resolve SlicerLiverInteractionLib on sys.path.
